jack: Fix upgrade path from jack1
The include order generated by the waf version packaged by jack2 is
incorrect, which leads to build failures when upgrading from jack1 to
jack2 because the build picks up the old jack headers. To avoid this,
add a version of the deactivate hack that will deactivate jack1 right
before the build.

Since this is an uncommon version of the deactivate hack that runs with
dropped privileges, separate privilege elevation is necessary.

Closes: https://trac.macports.org/ticket/56241